requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-celery-results==2.2.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djongo==1.3.6
dnspython==2.2.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
kombu==5.2.3
Markdown==3.3.6
MarkupSafe==2.1.0
openapi-codec==1.3.2
packaging==21.3
pika==1.2.0
prompt-toolkit==3.0.28
pydantic==1.9.0
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.7
pyrsistent==0.18.1
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.2.4
swagger-spec-validator==2.7.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
